requirements.txt
amqp==5.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2023.11.17
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
Flask==2.3.3
Flask-Caching==1.10.1
Flask-Limiter==1.4
gunicorn==20.1.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
limits==3.7.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
packaging==23.2
prompt-toolkit==3.0.43
protobuf==4.25.1
python-dotenv==1.0.0
pytz==2023.3.post1
redis==4.1.0
requests==2.27.1
six==1.16.0
typing_extensions==4.9.0
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
